Creación de clase PlataformaMovil y Palanca

+++ core/src/main/java/niveles/entorno/Plataforma.java
@@ -6,7 +6,7 @@ import com.badlogic.gdx.physics.box2d.World;
 
 public class Plataforma extends ElementoEntorno {
 
-    public Plataforma(World mundo, float x, float y, float ancho, float alto) {
+    public Plataforma(World mundo, float x, float y, int ancho, int alto) {
     	super(mundo, x, y, ancho, alto);
 
         textura = new Texture(Gdx.files.internal("plataforma.jpg"));
